// packages/mobile/src/shared/enums/validation-messages.enum.ts
export enum ValidationMessages {
	REQUIRED = "This field is required",
	INVALID_EMAIL = "Invalid email format",
	INVALID_PHONE = "Invalid phone number format",
	INVALID_PASSWORD = "Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character",
	PASSWORD_MIN_LENGTH = "Password must be at least 6 characters long",
	PASSWORDS_DO_NOT_MATCH = "Passwords do not match",
}